ch01: Greatly simplifying chapter one
References to vcpkg are no longer present in published book as I want to
focus on CMake only builds as I package managers are non-standard and
highly opinionated. FetchContent will suffice for the simple package
management examples that ch06 will feature.

The description of CMake presets has also been removed. While they are
useful and I prefer using them for configuring things like compiler flags,
cache variables etc. over directly in CMakeLists.txt files, they require
and additional file to be created and managed by readers and I wanted to
keep the opening chapter as concise as possible to allow for the reader
to move quickly to chapters 02 and 03. To compensate a more complex
CMake config will be covered in an appendix page with some in-between
content being featured in ch06 when discussing larger project
structures.

The whichlang JS and CSS files have been added to allow codeblocks to be
rendered in a more decorative manner. This was added to help codeblocks;
particularly for shell commands across platforms, to be more
distinguishable between each other. While PowerShell and Command Prompt
do not have icons due to a transient dependency; Simple Icons, not
having any the 'filepath' option allows for a text indication of which
shell a command is for.