ch2: Guessing Game migrated to new conventions and C++17
Previous source files used C++20 'Three-Way-Comparison' but this was not
required and kinda obfuscates how most code is written. Additionally,
with the (temporary) drop of C++20 as an implicit baseline in favour of
explicitly using C++17, TWC is no longer available and thus removed.

Styling of example code improved for readibility.

Instructions rewritten to reflect changed source.